Update the Metadata.json and screen shots for the Quartz Beta (Xamarin SDK's). Also renamed the sample: DisplayDrawStatus to DisplayDrawingStatus.
