--- src/Android/Xamarin.Android/Samples/MapView/DisplayDrawStatus/metadata.json
@@ -1,15 +1,0 @@
-{
-  "Name": "Display draw status",
-  "SampleName": "DisplayDrawStatus",
-  "Description": "This sample demonstrates how to display the draw status of a Map",
-  "Instructions": "",
-  "Type": 0,
-  "RequiresOnlineConnection":true,
-  "RequiresOfflineData":false,
-  "RequiresLocalServer":false,
-  "Image":"DisplayDrawStatus.jpg",
-  "Link":"",
-  "TypeLink": ["T:Esri.ArcGISRuntime.UI.DrawStatus","E:Esri.ArcGISRuntime.UI.GeoView.DrawStatusChanged"
-  ],
-  "SampleFolder": "DisplayDrawStatus"
-}
