+++ src/Forms/Shared/Samples/Map/DisplayMap/metadata.json
@@ -9,7 +9,7 @@
   "RequiresLocalServer":false,
   "Image":"DisplayMap.jpg",
   "Link":"",
-  "TypeLink": ["P:Esri.ArcGISRuntime.Mapping.Map.Basemap"
+  "TypeLink": ["T:Esri.ArcGISRuntime.Mapping.Basemap","T:Esri.ArcGISRuntime.Mapping.Map"
   ],
   "SampleFolder": "DisplayMap"
 }
